lavf/deinterlace_qsv: re-use VPPContext for deinterlace_qsv filter
All features are implemented in vpp_qsv filter now, so deinterlace_qsv
can be taken as a specical case of vpp_qsv filter, we re-use VPPContext
with a different option array and pix formats for deinterlace_qsv filter
